+++ docs/cell_images.py
@@ -92,8 +92,8 @@ def write_svg_image():
     # cell.write_svg(
     #     "example.svg",
     #     background="none",
-    #     style={(0, 1): {"fill": "none", "stroke": "black", "stroke-dasharray": "8,8"}},
-    #     fontstyle={(3, 2): {"fill": "none", "stroke": "red", "font-size": "32px"}},
+    #     shape_style={(0, 1): {"fill": "none", "stroke": "black", "stroke-dasharray": "8,8"}},
+    #     label_tyle={(3, 2): {"fill": "none", "stroke": "red", "font-size": "32px"}},
     #     pad="5%",
     #     sort_function=lambda p1, p2: p1.layer < p2.layer,
     # )