+++ docs/cpp/first.cpp
@@ -25,9 +25,7 @@ int main(int argc, char* argv[]) {
     lib.write_gds("first.gds", 0, NULL);
     lib.write_oas("first.oas", 0, 6, OASIS_CONFIG_DETECT_ALL);
 
-    StyleMap style = {0};
-    StyleMap label_style = {0};
-    cell.write_svg("first.svg", 10, 6, style, label_style, "#222222", 5, true, NULL);
+    cell.write_svg("first.svg", 10, 6, NULL, NULL, "#222222", 5, true, NULL);
 
     return 0;
 }