+++ docs/tutorial_images.py
@@ -19,8 +19,8 @@ def draw(cell, path):
         name,
         scaling=scaling,
         background="none",
-        style={(0, 1): {"fill": "none", "stroke": "black", "stroke-dasharray": "8,8"}},
-        fontstyle={(3, 2): {"stroke": "red", "fill": "none", "font-size": "32px"}},
+        shape_style={(0, 1): {"fill": "none", "stroke": "black", "stroke-dasharray": "8,8"}},
+        label_style={(3, 2): {"stroke": "red", "fill": "none", "font-size": "32px"}},
         pad="5%",
     )
     print(f"Saving {name} (scaling {scaling})")